# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mix REST API",
    "description": "REST API for the Mix application - session management, messaging, and system operations",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088",
      "description": "Development server"
    }
  ],
...
  "paths": {
    "/api/auth/api-key": {
      "post": {
        "description": "Store API key for direct authentication with a specific provider",
        "operationId": "storeApiKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "api_key": {
                    "description": "API key for authentication",
                    "type": "string"
                  },
                  "provider": {
                    "description": "Provider name (anthropic, openai, openrouter)",
                    "enum": [
                      "anthropic",
                      "openai",
                      "openrouter"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "provider",
                  "api_key"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "description": "Success message",
                      "type": "string"
                    },
                    "provider": {
                      "description": "Provider name",
                      "type": "string"
                    },
                    "status": {
                      "description": "Operation status (success)",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "API key stored status"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request data or API key format"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Store API key",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/apikey": {
      "post": {
        "description": "Set API key for direct authentication",
        "operationId": "setApiKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "apiKey": {
                    "description": "API key for authentication",
                    "type": "string"
                  }
                },
                "required": [
                  "apiKey"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "description": "Whether API key was set successfully",
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "API key set status"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid API key"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized - authentication failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Set API key",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "description": "Initiate OAuth authentication flow",
        "operationId": "initiateOAuthLogin",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "authUrl": {
                      "description": "OAuth authorization URL",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Authentication URL"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized - authentication failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "OAuth authentication",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/oauth-callback": {
      "post": {
        "description": "Process OAuth callback and exchange code for access token",
        "operationId": "handleOAuthCallback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "code": {
                    "description": "Authorization code from OAuth provider",
                    "type": "string"
                  },
                  "provider": {
                    "description": "Provider name (anthropic)",
                    "type": "string"
                  },
                  "state": {
                    "description": "OAuth state for verification",
                    "type": "string"
                  }
                },
                "required": [
                  "provider",
                  "code",
                  "state"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "expires_in": {
                      "description": "Seconds until token expiration",
                      "type": "integer"
                    },
                    "message": {
                      "description": "Status message",
                      "type": "string"
                    },
                    "provider": {
                      "description": "Provider name",
                      "type": "string"
                    },
                    "status": {
                      "description": "Operation status",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OAuth completion status"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request parameters"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Handle OAuth callback",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/oauth/{provider}": {
      "post": {
        "description": "Initiate OAuth authentication flow for a specific provider",
        "operationId": "startOAuthFlow",
        "parameters": [
          {
            "description": "Provider name (currently only 'anthropic')",
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "auth_url": {
                      "description": "OAuth authorization URL to redirect to",
                      "type": "string"
                    },
                    "message": {
                      "description": "Instructions for completing OAuth flow",
                      "type": "string"
                    },
                    "state": {
                      "description": "OAuth state token for verification",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OAuth authorization information"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid provider or OAuth not supported"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Start OAuth authentication",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/status": {
      "get": {
        "description": "Get authentication status for all supported providers",
        "operationId": "getAuthStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "providers": {
                      "additionalProperties": {
                        "properties": {
                          "auth_method": {
                            "description": "Authentication method (oauth, api_key, none)",
                            "enum": [
                              "oauth",
                              "api_key",
                              "none"
                            ],
                            "type": "string"
                          },
                          "authenticated": {
                            "description": "Whether provider is authenticated",
                            "type": "boolean"
                          },
                          "display_name": {
                            "description": "User-friendly provider name",
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "description": "Map of provider authentication status",
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Authentication status for all providers"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get authentication status",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/validate": {
      "get": {
        "description": "Check if the user's preferred provider is authenticated",
        "operationId": "validatePreferredProvider",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "auth_method": {
                      "description": "Authentication method used",
                      "enum": [
                        "oauth",
                        "api_key",
                        "none"
                      ],
                      "type": "string"
                    },
                    "message": {
                      "description": "Status message",
                      "type": "string"
                    },
                    "provider": {
                      "description": "Preferred provider name",
                      "type": "string"
                    },
                    "valid": {
                      "description": "Whether preferred provider is authenticated",
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Preferred provider validation status"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Validate preferred provider",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/{provider}": {
      "delete": {
        "description": "Delete stored API key and/or OAuth credentials for a provider",
        "operationId": "deleteCredentials",
        "parameters": [
          {
            "description": "Provider name (anthropic, openai, openrouter)",
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "description": "Status message",
                      "type": "string"
                    },
                    "provider": {
                      "description": "Provider name",
                      "type": "string"
                    },
                    "status": {
                      "description": "Operation status",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Credentials deletion status"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid provider"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Delete provider credentials",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/commands": {
      "get": {
        "description": "Retrieve list of all available commands",
        "operationId": "listCommands",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "description": {
                        "description": "Command description",
                        "type": "string"
                      },
                      "name": {
                        "description": "Command name",
                        "type": "string"
                      }
                    },
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of commands"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized - authentication required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "List available commands",
        "tags": [
          "System"
        ]
      }
    },
    "/api/commands/{name}": {
      "get": {
        "description": "Retrieve details about a specific command",
        "operationId": "getCommand",
        "parameters": [
          {
            "description": "Command name",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "description": {
                      "description": "Command description",
                      "type": "string"
                    },
                    "name": {
                      "description": "Command name",
                      "type": "string"
                    },
                    "usage": {
                      "description": "Command usage instructions",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Command details"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Command not found"
          }
        },
        "summary": "Get specific command",
        "tags": [
          "System"
        ]
      }
    },
    "/api/mcp": {
      "get": {
        "description": "Retrieve list of available Model Context Protocol (MCP) servers",
        "operationId": "listMcpServers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "connected": {
                        "description": "Whether the MCP server is currently connected",
                        "type": "boolean"
                      },
                      "name": {
                        "description": "MCP server name",
                        "type": "string"
                      },
                      "status": {
                        "description": "Server connection status (e.g., 'connected', 'failed', 'disconnected')",
                        "type": "string"
                      },
                      "tools": {
                        "description": "List of tools provided by this MCP server (null if server is not connected)",
                        "items": {
                          "properties": {
                            "description": {
                              "description": "Tool description",
                              "type": "string"
                            },
                            "name": {
                              "description": "Tool name",
                              "type": "string"
                            }
                          },
                          "required": [
                            "name",
                            "description"
                          ],
                          "type": "object"
                        },
                        "nullable": true,
                        "type": "array"
                      }
                    },
                    "required": [
                      "name",
                      "connected",
                      "status"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of MCP servers"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized - authentication required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "List MCP servers",
        "tags": [
          "System"
        ]
      }
    },
    "/api/messages/history": {
      "get": {
        "description": "Retrieve message history across all sessions with optional pagination",
        "operationId": "getMessageHistory",
        "parameters": [
          {
            "description": "Maximum number of messages to return",
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 50,
              "maximum": 1000,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Number of messages to skip",
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageData"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Message history"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized - authentication required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get global message history",
        "tags": [
          "Messages"
        ]
      }
    },
    "/api/permissions/{id}/deny": {
      "post": {
        "description": "Deny a specific permission",
        "operationId": "denyPermission",
        "parameters": [
          {
            "description": "Permission ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "denied": {
                      "description": "Whether permission was denied",
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Permission deny status"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized - authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Permission not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Deny permission",
        "tags": [
          "Permissions"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/RESTError"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "MessageData": {
        "properties": {
          "assistantResponse": {
            "description": "Assistant's response message (optional)",
            "type": "string"
          },
          "id": {
            "description": "Unique message identifier",
            "type": "string"
          },
          "reasoning": {
            "description": "Reasoning process (optional)",
            "type": "string"
          },
          "reasoningDuration": {
            "description": "Reasoning duration in milliseconds (optional)",
            "type": "integer"
          },
          "role": {
            "$ref": "#/components/schemas/MessageRole"
          },
          "sessionId": {
            "description": "Session identifier",
            "type": "string"
          },
          "toolCalls": {
            "description": "Tool calls made during message processing",
            "items": {
              "$ref": "#/components/schemas/ToolCallData"
            },
            "type": "array"
          },
          "userInput": {
            "description": "User's input message",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionId",
          "role",
          "userInput"
        ],
        "type": "object"
      },
      "MessageRole": {
        "description": "Message role",
        "enum": [
          "user",
          "assistant",
          "tool"
        ],
        "type": "string"
      },
      "RESTError": {
        "properties": {
          "code": {
            "description": "HTTP status code",
            "type": "integer"
          },
          "message": {
            "description": "Error message",
            "type": "string"
          },
          "type": {
            "description": "Error type",
            "enum": [
              "bad_request",
              "not_found",
              "internal_error",
              "unauthorized",
              "validation_error"
            ],
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "type"
        ],
        "type": "object"
      },
...
    }
  }
}